+++ src/jinnang/__init__.py
@@ -1,7 +1,7 @@
 # Utils module
 
 # Common utilities
-from jinnang.common.formatting import safe_format, date_str_to_iso_date_str, get_int, timestamp_to_date
+from jinnang.common.formatting_utils import safe_format, date_str_to_iso_date_str, get_int, timestamp_to_date
 
 # IO and System utilities
 from jinnang.io.system import suppress_stdout_stderr, get_worker_num_for_io_bounded_task, safe_delete, safe_move, copy_with_meta
@@ -12,7 +12,9 @@ from jinnang.data.geo_utils import calculate_distance_meters
 # Common utilities
 from jinnang.common.decorators import mock_when, fail_recover, custom_retry
 from jinnang.common.exceptions.exceptions import BadInputException
-from jinnang.common.formatters import ResolutionPreset, Verbosity, TruncatedPrettyPrinter, get_numeric
+from jinnang.common.formatters import get_numeric
+from jinnang.common.verbosity import Verbosity
+from jinnang.media.resolution import ResolutionPreset
 from jinnang.common.path import MyPath, get_file_timestamp, get_video_duration
 
 # Text utilities