+++ src/jinnang/common/__init__.py
@@ -2,8 +2,9 @@
 from jinnang.common.collections import get_first_value, get_unique_key
 from jinnang.common.decorators import mock_when, fail_recover, custom_retry
 from jinnang.common.exceptions.exceptions import BadInputException
-from jinnang.common.formatters import ResolutionPreset, Verbosity, TruncatedPrettyPrinter, get_numeric
+from jinnang.common.formatters import get_numeric
+from jinnang.common.verbosity import Verbosity
 from jinnang.common.path import MyPath, ensure_unique_path, get_file_timestamp, timestamp_to_date, get_video_duration
 from jinnang.common.patterns import GenericSingletonFactory
 from jinnang.common.arithmetic import *
-from jinnang.common.formatting import *
+from jinnang.common.formatting_utils import *